+++ library/src/methods/safeParse/safeParse.ts
@@ -1,5 +1,6 @@
-import { type Issues, ValiError } from '../../error/index.ts';
-import type { BaseSchema, Output, ParseInfo } from '../../types.ts';
+import { ValiError } from '../../error/index.ts';
+import type { BaseSchema, ParseInfo } from '../../types.ts';
+import type { SafeParseResult } from './types.ts';
 
 /**
  * Parses unknown input based on a schema.
@@ -14,16 +15,7 @@ export function safeParse<TSchema extends BaseSchema>(
   schema: TSchema,
   input: unknown,
   info?: Pick<ParseInfo, 'abortEarly' | 'abortPipeEarly'>
-):
-  | { success: true; data: Output<TSchema> }
-  | {
-      success: false;
-      /**
-       * @deprecated Please use `.issues` instead.
-       */
-      error: ValiError;
-      issues: Issues;
-    } {
+): SafeParseResult<TSchema> {
   const result = schema._parse(input, info);
   return result.issues
     ? {
@@ -31,5 +23,9 @@ export function safeParse<TSchema extends BaseSchema>(
         error: new ValiError(result.issues),
         issues: result.issues,
       }
-    : { success: true, data: result.output };
+    : {
+        success: true,
+        data: result.output,
+        output: result.output,
+      };
 }