+++ library/src/schemas/literal/types.ts
@@ -1,0 +1,4 @@
+/**
+ * Literal type.
+ */
+export type Literal = number | string | boolean | symbol | bigint;