+++ library/src/transformations/toLowerCase/toLowerCase.ts
@@ -1,4 +1,4 @@
-import type { ParseResult } from '../../types.ts';
+import type { _ParseResult } from '../../types.ts';
 
 /**
  * Creates a transformation function that converts all the alphabetic
@@ -7,7 +7,7 @@ import type { ParseResult } from '../../types.ts';
  * @returns A transformation function.
  */
 export function toLowerCase() {
-  return (input: string): ParseResult<string> => ({
+  return (input: string): _ParseResult<string> => ({
     output: input.toLocaleLowerCase(),
   });
 }