+++ library/src/validations/includes/includes.ts
@@ -1,15 +1,15 @@
-import type { ParseResult, ValidateInfo } from '../../types.ts';
+import type { _ParseResult, ValidateInfo } from '../../types.ts';
 import { getIssue } from '../../utils/index.ts';
 
 export function includes<TInput extends string>(
   requirement: string,
   error?: string
-): (input: TInput, info: ValidateInfo) => ParseResult<TInput>;
+): (input: TInput, info: ValidateInfo) => _ParseResult<TInput>;
 
 export function includes<TInput extends TItem[], TItem>(
   requirement: TItem,
   error?: string
-): (input: TInput, info: ValidateInfo) => ParseResult<TInput>;
+): (input: TInput, info: ValidateInfo) => _ParseResult<TInput>;
 
 /**
  * Creates a validation functions that validates the content of a string or array.
@@ -23,7 +23,7 @@ export function includes<TInput extends string | TItem[], TItem>(
   requirement: string | TItem,
   error?: string
 ) {
-  return (input: TInput, info: ValidateInfo): ParseResult<TInput> => {
+  return (input: TInput, info: ValidateInfo): _ParseResult<TInput> => {
     if (!input.includes(requirement as any)) {
       return {
         issues: [